删除栏目修改为 当栏目内文章不为空时无法删除

+++ routes/category/deleteCategory.js
@@ -2,7 +2,9 @@ const { resTemp, errTemp } = require('../config');
 const { CategoryModel } = require('../../db');
 
 module.exports = (req, res) => {
-  const { categoryId } = req.body;
+  const { categoryId, articleCount } = req.body;
+  if (articleCount > 0)
+    return res.json(errTemp({}, '该栏目不为空，无法删除！'));
   if (!!categoryId) {
     CategoryModel.findByIdAndDelete(categoryId, (err, category) => {
       if (!!err) res.json(errTemp(err, ''));
